+++ CHANGELOG.md
@@ -13,6 +13,9 @@
     - menu: merge different interaction modes into a single `mode` property
     - menu: add `swipe` mode
 
+- remote control
+    - `/STATE/SET` add optional argument to prevent widgets from sending their value
+
 ## 1.24.2
 
 - bug fixes